Just playing around testing my idea
I had the idea of creating a PropTypes like validation scheme for php. I
thought it might be very useful for validating JSON. I know I could use
JSON-Schema but I think working with it is a pain in the butt. I like
how PropTypes is keeping the structure of the validation scheme the same
as any objct you would validate with it while at the same time making
validation quite uncomplicated.