Added object structure validation
To make this system very useful and able to work with JSON data at all,
I need to be able to validate objects. With this I added this
functionality. All very basic, and I can already see some flaws in it.
I can't add functionality that checks for empy objects, like the
`nonEmpty` validator for strings. I can not ommit the content of the
object in case I just want to make sure an object is there, without
checking its structure.
I'll address those in the coming commits.